refactor: Require request access for all communities
This commit introduces a few changes regarding users accessing
communities:

- While the APIs still exist, community invites should no longer be
used, instead communities should merely be "shared".
- Sharing a community to users allows users to "join" the community,
which in reality makes them request access to that community.
- This means, users have to request access to any community, even if
the community has permissions set to NO_MEMBERSHIP
- Only difference between ON_REQUEST and NO_MEMBERSHIP is that
ON_REQUEST communities require manual approval of the owner/admin
to access a community. NO_MEMBERSHIP communities accept
automatically (as soon as owner/admin receives the request).
- Nodes then have to wait for a `CommunityRequestToJoinResponse` to know
wether they have been accepted/joined or not.

This also implies that users are no longer optimistically added to the
member list of communities, but only after they have been accepted.

This introduces a bit of a message ping-pong for users to know that
someone is now part of a community.